+++ packages/dialtone-emojis/release-local.config.cjs
@@ -2,27 +2,31 @@
 const name = 'dialtone-emojis';
 const srcRoot = `packages/${name}`;
 
+/**
+ * @type {import('semantic-release').GlobalConfig}
+ */
 module.exports = {
   extends: 'release.config.base.js',
+  ci: false,
   pkgRoot: srcRoot,
   tagFormat: name + '/v${version}',
   commitPaths: [`${srcRoot}/*`],
   assets: [
     `${srcRoot}/CHANGELOG.md`,
     `${srcRoot}/CHANGELOG.json`,
-    `${srcRoot}/package.json`
+    `${srcRoot}/package.json`,
   ],
   plugins: [
     ['@semantic-release/commit-analyzer', {
       preset: 'angular',
       releaseRules: [
-        {type: 'build', release: 'patch'},
-        {type: 'chore', release: 'patch'},
-        {type: 'ci', release: 'patch'},
-        {type: 'docs', release: 'patch'},
-        {type: 'refactor', release: 'patch'},
-        {type: 'style', release: 'patch'},
-        {type: 'test', release: 'patch'},
+        { type: 'build', release: 'patch' },
+        { type: 'chore', release: 'patch' },
+        { type: 'ci', release: 'patch' },
+        { type: 'docs', release: 'patch' },
+        { type: 'refactor', release: 'patch' },
+        { type: 'style', release: 'patch' },
+        { type: 'test', release: 'patch' },
       ],
     }],
     ['@semantic-release/release-notes-generator', {
@@ -35,7 +39,6 @@ module.exports = {
     ['@semantic-release/changelog', { changelogFile: `${srcRoot}/CHANGELOG.md` }],
     ['@semantic-release/npm', { npmPublish: false }],
     ['@semantic-release/git', {
-      /* eslint-disable-next-line no-template-curly-in-string */
       message: `chore(release): NO-JIRA ${name}` +
         '/v${nextRelease.version}\n\n${nextRelease.notes}',
     }],