+++ packages/dialtone-icons/vue3/vite.config.js
@@ -23,7 +23,6 @@ export default defineConfig({
       external: ['vue'],
       output: {
         minifyInternalExports: true,
-        sourcemap: true,
       },
     },
     minify: true,