+++ packages/stylelint-plugin-dialtone/release-ci.config.cjs
@@ -1,6 +1,10 @@
+/* eslint-disable no-template-curly-in-string */
 const name = 'stylelint-plugin-dialtone';
 const srcRoot = `packages/${name}`;
 
+/**
+ * @type {import('semantic-release').GlobalConfig}
+ */
 module.exports = {
   extends: 'release.config.base.js',
   pkgRoot: srcRoot,