+++ release-ci.config.cjs
@@ -2,6 +2,9 @@
 const name = 'dialtone';
 const srcRoot = '.';
 
+/**
+ * @type {import('semantic-release').GlobalConfig}
+ */
 module.exports = {
   extends: 'release.config.base.js',
   pkgRoot: srcRoot,